Use subprocess instead of multiprocessing for test_common_run_task.py.
